cleaned up the code and added gnss_connection header and source files
