+++ Kconfig
@@ -1,9 +1,3 @@
-#
-# Copyright (c) 2020 Nordic Semiconductor ASA
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-#
-
 menu "FishIoT"
 
 config MQTT_PUB_TOPIC
@@ -82,7 +76,7 @@ config GNSS_PERIODIC_INTERVAL
 config GNSS_PERIODIC_TIMEOUT
 	int "Fix timeout for periodic GPS fixes"
 	range 0 65535
-	default 480
+	default 180
 	help
 	  Fix timeout (in seconds) for periodic fixes.
 	  If set to zero, GNSS is allowed to run indefinitely until a valid PVT estimate is produced.
