+++ prj.conf
@@ -1,9 +1,3 @@
-#
-# Copyright (c) 2020 Nordic Semiconductor ASA
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-#
-
 # Logging
 CONFIG_LOG=y
 
@@ -30,24 +24,24 @@ CONFIG_NRF_MODEM_LIB=y
 CONFIG_NRF_MODEM_LIB_SYS_INIT=n
 
 # Enable the nRF9160 modem key management library
-CONFIG_MODEM_KEY_MGMT=y
+# CONFIG_MODEM_KEY_MGMT=y
 # LTE link control
 CONFIG_LTE_LINK_CONTROL=y
 # CONFIG_LTE_AUTO_INIT_AND_CONNECT is deprecated, and kept for compatibility with older NCS versions
 CONFIG_LTE_AUTO_INIT_AND_CONNECT=n
 
-# MQTT
-CONFIG_MQTT_LIB=y
-CONFIG_MQTT_CLEAN_SESSION=y
-# Enable TLS for the MQTT library
-CONFIG_MQTT_LIB_TLS=y
-# Application
-CONFIG_MQTT_PUB_TOPIC="test"
-CONFIG_MQTT_SUB_TOPIC="test"
-CONFIG_BUTTON_EVENT_PUBLISH_MSG="Hi from the nRF9160 SiP"
-CONFIG_MQTT_BROKER_HOSTNAME="34.88.119.114"
-# Change the MQTT broker port
-CONFIG_MQTT_BROKER_PORT=8883
+# # MQTT
+# CONFIG_MQTT_LIB=y
+# CONFIG_MQTT_CLEAN_SESSION=y
+# # Enable TLS for the MQTT library
+# CONFIG_MQTT_LIB_TLS=y
+# # Application
+# CONFIG_MQTT_PUB_TOPIC="test"
+# CONFIG_MQTT_SUB_TOPIC="test"
+# CONFIG_BUTTON_EVENT_PUBLISH_MSG="Hi from the nRF9160 SiP"
+# CONFIG_MQTT_BROKER_HOSTNAME="34.88.119.114"
+# # Change the MQTT broker port
+# CONFIG_MQTT_BROKER_PORT=8883
 
 # Request PSM periodic TAU and active time
 CONFIG_LTE_PSM_REQ_RPTAU="00100010" #2hours
@@ -71,12 +65,9 @@ CONFIG_MODEM_JWT=y
 CONFIG_NRF_CLOUD_REST_AUTOGEN_JWT=n
 
 CONFIG_NRF_CLOUD_CLIENT_ID_SRC_IMEI=y
-# CONFIG_NRF_CLOUD_CONNECTION_POLL_THREAD=y
-# CONFIG_NRF_CLOUD_CONNECTION_POLL_THREAD_STACK_SIZE=4096
 
 CONFIG_NRF_CLOUD_GPS_LOG_LEVEL_DBG=y
 CONFIG_NRF_CLOUD_LOG_LEVEL_DBG=y
 CONFIG_DATE_TIME=y
 CONFIG_DATE_TIME_MODEM=y
-CONFIG_DATE_TIME_AUTO_UPDATE=n
 CONFIG_DATE_TIME_NTP=y
